[Siddartha]:Refactor:Refactored Asset Controller and Service

+++ src/main/java/com/eams/service/AssetService.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 
 import com.eams.dtos.AssetDTO;
 import com.eams.entity.Asset;
+import com.eams.entity.Role;
 import com.eams.entity.User;
 import com.eams.repository.AssetRepository;
 import com.eams.repository.UserRepository;
@@ -19,20 +20,30 @@ public class AssetService {
 	
 	@Autowired 
 	private UserRepository userRepository;
-	//TODO : try to give return type a boolean or any response message of the status rather than returning whole entity back.
-	public Asset createAsset(AssetDTO dto) {
-		
-		User user =userRepository.findById(dto.getAssignedTo()).orElseThrow();
-		Asset asset = Asset.builder()
-				.asset_name(dto.getAsset_name())
-				.asset_type(dto.getAsset_type())
-				.location(dto.getLocation())
-				.thresholdTemp(dto.getThresholdTemp())
-				.thresholdPressure(dto.getThresholdPressure())
-				.assignedTo(user)
-				.build();
-		return assetRepository.save(asset);
+
+	public boolean createAsset(AssetDTO dto) {
+	    try {
+	        User user = userRepository.findById(dto.getAssignedTo())
+	                .orElseThrow();
+	        if(user.getRole()==Role.MANAGER) {
+	        Asset asset = Asset.builder()
+	                .asset_name(dto.getAsset_name())
+	                .asset_type(dto.getAsset_type())
+	                .location(dto.getLocation())
+	                .thresholdTemp(dto.getThresholdTemp())
+	                .thresholdPressure(dto.getThresholdPressure())
+	                .assignedTo(user)
+	                .build();
+
+	        assetRepository.save(asset);
+	        return true;	        
+	        }
+	    } catch (Exception e) {
+	        System.out.println("Role is not Manager "+e.getMessage());
+	    }
+	    return false;
 	}
+
 	
 	public List<Asset> getAllAssets() {
 		return assetRepository.findAll();
@@ -42,25 +53,53 @@ public class AssetService {
 	public Asset getAssetById(Long id) {
 		return assetRepository.findById(id).orElseThrow();
 	}
-	//TODO: use try catch block or throws incase given id is not present.
-	public void deleteAsset(Long id) {
-		assetRepository.deleteById(id);
-	}
-	//TODO: use try catch block or throws incase given id is not present.
-	public Asset updateAsset(Long id,AssetDTO dto) {
-		Asset asset = assetRepository.findById(id).orElseThrow();
-		User user = userRepository.findById(dto.getAssignedTo()).orElseThrow();
+	
+
+	public String deleteAsset(Long id,Long userid) {
 		
-		asset.setAsset_name(dto.getAsset_name());
-		asset.setAsset_type(dto.getAsset_type());
-		asset.setLocation(dto.getLocation());
-		asset.setThresholdTemp(dto.getThresholdTemp());
-		asset.setThresholdPressure(dto.getThresholdPressure());
-		asset.setAssignedTo(user);
-		return assetRepository.save(asset);
-		
+	    try {
+	    	 User user = userRepository.findById(userid)
+		                .orElseThrow();
+		    if(user.getRole()==Role.MANAGER) {
+	        Asset as = assetRepository.findById(id)
+	                .orElseThrow();
+	        
+	        assetRepository.deleteById(id);
+	        return "Asset deleted successfully.";
+		    }
+		    else {
+		    	return "Only Manager can delete assests";
+		    }
+	    }  catch (Exception ex) {
+	        return "Error deleting asset: " + ex.getMessage();
+	    }
 	}
 	
+
+	public String updateAsset(Long id, AssetDTO dto) {
+	    try {
+	        Asset asset = assetRepository.findById(id)
+	                .orElseThrow();
+
+	        User user = userRepository.findById(dto.getAssignedTo())
+	                .orElseThrow();
+	
+	        asset.setAsset_name(dto.getAsset_name());
+	        asset.setAsset_type(dto.getAsset_type());
+	        asset.setLocation(dto.getLocation());
+	        asset.setThresholdTemp(dto.getThresholdTemp());
+	        asset.setThresholdPressure(dto.getThresholdPressure());
+	        asset.setAssignedTo(user);
+
+	        assetRepository.save(asset);
+	        return "Asset updated successfully.";
+	        
+	    }  catch (Exception ex) {
+	        return "Error updating asset: " + ex.getMessage();
+	    }
+	}
+
+
